requirements.txt
beautifulsoup4==4.12.2
functions_framework==3.3.0
google_api_python_client==2.86.0
google_auth_oauthlib==1.0.0
protobuf==4.23.1
selenium==4.9.1
webdriver_manager==3.8.6
